+++ lib/src/style.dart
@@ -742,6 +742,9 @@ class TxtStyle extends StyleClass {
 
   void wordSpacing(double space) => _textModel?.wordSpacing = space;
 
+  void textDecoration(TextDecoration decoration) =>
+      _textModel?.textDecoration = decoration;
+
   /// Make the widget editable just like a TextField.
   ///
   /// If `focusNode` isnt spesified an internal `focusNode` will be created.
@@ -749,18 +752,22 @@ class TxtStyle extends StyleClass {
       {TextInputType keyboardType,
       String placeholder,
       bool obscureText = false,
+      int maxLines,
       void Function(String) onChange,
       void Function(bool focus) onFocusChange,
       void Function(TextSelection, SelectionChangedCause) onSelectionChanged,
+      void Function() onEditingComplete,
       FocusNode focusNode}) {
     _textModel
       ..editable = enable
       ..keyboardType = keyboardType
       ..placeholder = placeholder
       ..obscureText = obscureText
+      ..maxLines = maxLines
       ..onChange = onChange
       ..onFocusChange = onFocusChange
       ..onSelectionChanged = onSelectionChanged
+      ..onEditingComplete = onEditingComplete
       ..focusNode = focusNode;
   }